Add files via upload
This file contains the preprocessing from the original skip_model.ipynb file as well as the model from the toy_model.ipynb file. The key changes in this new skip_model_v2.ipynb file are in regards to the variables used in the model. This new model uses more song and listening session level characteristics that are not included in the toy_model.ipynb file.